hakaton/requirements.txt
Django==3.2.3
requests==2.26.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
django-filter==23.2
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
python-dotenv==1.0.0
gunicorn==20.1.0
drf-extra-fields==3.4.0
openpyxl==3.1.2
pandas==2.1.1
xlwt==1.3.0
django-excel-response==2.0.5